Safari-611.1.10/LayoutTests/platform/mac-mojave/http/tests/cookies/same-site/popup-cross-site-post-expected.txt


FAIL '127.0.0.1' is not same-site with 'localhost', so samesite cookies are not sent via POST. assert_equals: implicit-strict expected (string) "1" but got (undefined) undefined
